Renamed Symantics to ClassInterface, split out GOOL's internal classes for renderers to RendererClasses
