+++ code/drasil-code/Language/Drasil/Code/Imperative/GOOL/ClassInterface.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE TypeFamilies #-}
 
-module Language.Drasil.Code.Imperative.GOOL.Symantics (
+module Language.Drasil.Code.Imperative.GOOL.ClassInterface (
   -- Typeclasses
   PackageSym(..), AuxiliarySym(..)
 ) where
