+++ code/drasil-code/Language/Drasil/Code/Imperative/GOOL/LanguageRenderer/CppRenderer.hs
@@ -8,7 +8,7 @@ module Language.Drasil.Code.Imperative.GOOL.LanguageRenderer.CppRenderer (
 ) where
 
 import Language.Drasil.CodeSpec (ImplementationType(..))
-import Language.Drasil.Code.Imperative.GOOL.Symantics (PackageSym(..),
+import Language.Drasil.Code.Imperative.GOOL.ClassInterface (PackageSym(..),
   AuxiliarySym(..))
 import qualified 
   Language.Drasil.Code.Imperative.GOOL.LanguageRenderer.LanguagePolymorphic as
