+++ code/drasil-gool/GOOL/Drasil/LanguageRenderer/LanguagePolymorphic.hs
@@ -42,29 +42,19 @@ module GOOL.Drasil.LanguageRenderer.LanguagePolymorphic (fileFromData, oneLiner,
 import Utils.Drasil (indent)
 
 import GOOL.Drasil.CodeType (CodeType(..), ClassName)
-import GOOL.Drasil.Symantics (Label, Library, KeywordSym(..), RenderSym,
+import GOOL.Drasil.ClassInterface (Label, Library, 
   FileSym(RenderFile, commentedMod), BodySym(Body, body, bodyStatements), 
-  InternalBody(bodyDoc, docBody), BlockSym(Block), 
-  InternalBlock(docBlock, blockDoc), ImportSym(..), 
-  PermanenceSym(..), InternalPerm(..), 
+  BlockSym(Block), PermanenceSym(..), 
   TypeSym(Type, infile, outfile, iterator, getType, getTypeDoc, getTypeString), 
-  InternalType(..), UnaryOpSym(UnaryOp), BinaryOpSym(BinaryOp), InternalOp(..),
   VariableSym(Variable, variableName, variableType), 
-  InternalVariable(variableBind, variableDoc, varFromData), 
   ValueSym(Value, valueDoc, valueType), 
   NumericExpression((#+), (#-), (#*), (#/), sin, cos, tan), 
-  BooleanExpression(..), InternalValue(inputFunc, cast, valuePrec, valFromData),
-  Selector(($.)), FunctionSym(Function), SelectorFunction(at), 
-  InternalFunction(iterBeginFunc, iterEndFunc, functionDoc, functionType, 
-    funcFromData),
-  InternalStatement(statementDoc, statementTerm, stateFromData), 
+  BooleanExpression(..), Selector(($.)), FunctionSym(Function), 
+  SelectorFunction(at), 
   StatementSym(Statement, (&=), (&+=), (&++), break, multi), ScopeSym(..),
-  InternalScope(..), MethodTypeSym(MethodType, mType), ParameterSym(Parameter), 
-  InternalParam(paramFromData), MethodSym(Method), 
-  InternalMethod(intMethod, commentedFunc, methodDoc), StateVarSym(StateVar), 
-  InternalStateVar(..), ClassSym(Class), InternalClass(classDoc, classFromData),
-  ModuleSym(Module), InternalMod(moduleDoc, updateModuleDoc), BlockComment(..))
-import qualified GOOL.Drasil.Symantics as S (InternalFile(fileFromData), 
+  ParameterSym(Parameter), MethodSym(Method), StateVarSym(StateVar), 
+  ClassSym(Class), ModuleSym(Module), BlockComment(..))
+import qualified GOOL.Drasil.ClassInterface as S (
   BodySym(oneLiner), BlockSym(block), 
   TypeSym(bool, int, float, double, char, string, listType, arrayType, 
     listInnerType, void), 
@@ -73,18 +63,33 @@ import qualified GOOL.Drasil.Symantics as S (InternalFile(fileFromData),
   ValueExpression(funcApp, funcAppMixedArgs, selfFuncAppMixedArgs, 
     extFuncAppMixedArgs, libFuncAppMixedArgs, newObj, newObjMixedArgs, 
     extNewObj, extNewObjMixedArgs, libNewObjMixedArgs, notNull, lambda), 
-  InternalValue(call), Selector(objAccess), objMethodCall, 
+  Selector(objAccess), objMethodCall, 
   objMethodCallNoParams, FunctionSym(func, listSize, listAdd, listAppend),
   SelectorFunction(listAccess, listSet),
-  InternalFunction(getFunc, setFunc, listSizeFunc, listAddFunc, listAppendFunc, 
-    listAccessFunc, listSetFunc),
-  InternalStatement(state, loopState, emptyState), 
   StatementSym(assign, varDec, varDecDef, listDec, listDecDef, objDecNew, 
     extObjDecNew, constDecDef, valState, returnState),
-  ControlStatementSym(ifCond, for, forRange, switch), MethodTypeSym(construct), 
-  ParameterSym(param), MethodSym(method, mainFunction), InternalMethod(intFunc),
-  StateVarSym(stateVar), ClassSym(buildClass, commentedClass), 
-  InternalClass(intClass), InternalMod(modFromData))
+  ControlStatementSym(ifCond, for, forRange, switch),
+  ParameterSym(param), MethodSym(method, mainFunction),
+  StateVarSym(stateVar), ClassSym(buildClass, commentedClass))
+import GOOL.Drasil.RendererClasses (KeywordSym(..), RenderSym, 
+  InternalBody(bodyDoc, docBody), InternalBlock(docBlock, blockDoc), 
+  ImportSym(..), InternalPerm(..), InternalType(..), UnaryOpSym(UnaryOp), 
+  BinaryOpSym(BinaryOp), InternalOp(..), 
+  InternalVariable(variableBind, variableDoc, varFromData), 
+  InternalValue(inputFunc, cast, valuePrec, valFromData),
+  InternalFunction(iterBeginFunc, iterEndFunc, functionDoc, functionType, 
+    funcFromData),
+  InternalStatement(statementDoc, statementTerm, stateFromData), 
+  InternalScope(..), MethodTypeSym(MethodType, mType), 
+  InternalParam(paramFromData), 
+  InternalMethod(intMethod, commentedFunc, methodDoc), InternalStateVar(..), 
+  InternalClass(classDoc, classFromData), 
+  InternalMod(moduleDoc, updateModuleDoc))
+import qualified GOOL.Drasil.RendererClasses as S (InternalFile(fileFromData), 
+  InternalValue(call), InternalFunction(getFunc, setFunc, listSizeFunc, 
+    listAddFunc, listAppendFunc, listAccessFunc, listSetFunc),
+  InternalStatement(state, loopState, emptyState), MethodTypeSym(construct), 
+  InternalMethod(intFunc), InternalClass(intClass), InternalMod(modFromData))
 import GOOL.Drasil.AST (Binding(..), ScopeTag(..), Terminator(..), isSource)
 import GOOL.Drasil.Helpers (angles, doubleQuotedText, vibcat, emptyIfEmpty, 
   toState, onStateValue, on2StateValues, on3StateValues, onStateList,
